chore: clean up code style and split SolidSyslog.h (ISP)
Split SolidSyslog.h into two audience-specific headers:
- SolidSyslog.h — logging API (SolidSyslogMessage, SolidSyslog_Log)
- SolidSyslogConfig.h — lifecycle (SolidSyslogConfig, Create, Destroy)

Clean up SolidSyslogUdpSender.c: reorder to Create/Destroy first,
extract ResolveAddress and BuildAddress helpers, forward declarations
at top alphabetically.

Clean up SolidSyslogUdpSenderTest.cpp: add Send overloads to test
groups, removing direct SolidSyslogSender_Send calls from test bodies.

Prune implemented items from IGNORE_TEST todo comments.

Update CLAUDE.md and README.md with interface audience guidance.
