+++ Interface/SolidSyslog.h
@@ -1,20 +1,12 @@
 #ifndef SOLIDSYSLOG_H
 #define SOLIDSYSLOG_H
 
-#include "SolidSyslogAlloc.h"
+#include "ExternC.h"
 #include "SolidSyslogPrival.h"
 
 EXTERN_C_BEGIN
 
-    struct SolidSyslogSender;
     struct SolidSyslog;
-
-    struct SolidSyslogConfig
-    {
-        struct SolidSyslogSender* sender;
-        SolidSyslogAllocFn        alloc;
-        SolidSyslogFreeFn         free;
-    };
 
     struct SolidSyslogMessage
     {
@@ -22,9 +14,7 @@ EXTERN_C_BEGIN
         enum SolidSyslog_Severity severity;
     };
 
-    struct SolidSyslog* SolidSyslog_Create(const struct SolidSyslogConfig* config);
-    void                SolidSyslog_Destroy(struct SolidSyslog * logger);
-    void                SolidSyslog_Log(struct SolidSyslog * logger, const struct SolidSyslogMessage* message);
+    void SolidSyslog_Log(struct SolidSyslog * logger, const struct SolidSyslogMessage* message);
 
 EXTERN_C_END
